+++ emulator/opcode6.go
@@ -2,9 +2,9 @@ package emulator
 
 // RTSImplied 0x60: Return from Subroutine
 func (cpu *CPU) RTSImplied() {
-	lower := uint16(cpu.FetchMemory8((0x100 + uint(cpu.Reg.S) - 1)))
+	lower := uint16(cpu.FetchMemory8((0x100 + uint16(cpu.Reg.S) - 1)))
 	cpu.Reg.S--
-	upper := uint16(cpu.FetchMemory8((0x100 + uint(cpu.Reg.S) - 1)))
+	upper := uint16(cpu.FetchMemory8((0x100 + uint16(cpu.Reg.S) - 1)))
 	cpu.Reg.S--
 	cpu.Reg.PC = (upper << 8) | lower
 	cpu.Reg.PC++
@@ -30,7 +30,7 @@ func (cpu *CPU) RORZeroPage() {
 
 // PLAImplied 0x68: Pull A from stack (stack -> A)
 func (cpu *CPU) PLAImplied() {
-	value := cpu.FetchMemory8(0x0100 + uint(cpu.Reg.S) - 1)
+	value := cpu.FetchMemory8(0x0100 + uint16(cpu.Reg.S) - 1)
 	cpu.Reg.A = value
 	cpu.Reg.S--
 	cpu.Reg.PC++
