+++ emulator/opcode9.go
@@ -6,7 +6,7 @@ func (cpu *CPU) BCCRelative() {
 
 	cFlag := uint8(cpu.Reg.P & 0x01)
 	if cFlag == 0 {
-		cpu.Reg.PC = uint16(addr)
+		cpu.Reg.PC = addr
 	}
 }
